fix all clippy warnings and diagnostics
- Add crate-level #![allow(dead_code)] for API structs/functions used via serde/frontend
- Prefix unused variables with underscore
- Add #[allow(clippy::large_enum_variant)] for WebSocketEvent and CacheResult
- Add #[allow(clippy::type_complexity)] for complex type signatures
- Add #[allow(clippy::too_many_arguments)] for functions with many params
- Replace match with matches! macro in is_source_id
- Replace sort_by with sort_by_key using Reverse
- Simplify match with unwrap_or_default
- Replace .clone() with std::slice::from_ref

+++ src/embeddings.rs
@@ -336,7 +336,7 @@ pub async fn compute_tag_similarities(pool: &sqlx::Pool<sqlx::Sqlite>) -> Result
                 )
                 .bind(tag_ids[i])
                 .bind(tag_ids[j])
-                .bind(sim as f32)
+                .bind(sim)
                 .bind(now)
                 .execute(&mut *tx)
                 .await?;
